# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "contact": {}
  },
  "paths": {
    "/": {
      "get": {
        "description": "Redirect to origin URL by short URL",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/html; charset=utf-8"
        ],
        "operationId": "redirect",
        "responses": {
          "307": {
            "description": "Temporary Redirect",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "integer"
            }
          }
        }
      },
      "post": {
        "description": "Short URL",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/plain"
        ],
        "operationId": "shorten",
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    },
    "/api/shorten": {
      "post": {
        "description": "Short URL in json format",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "shorten-json",
        "parameters": [
          {
            "description": "URL for shortening",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.ShortenInput"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    },
    "/api/shorten/batch": {
      "post": {
        "description": "Short list of URLs in json format",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "shorten-batch-json",
        "parameters": [
          {
            "description": "URL for shortening",
            "name": "input",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "correlation_id": {
                    "type": "string"
                  },
                  "original_url": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "307": {
            "description": "Temporary Redirect",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    },
    "/api/user/urls": {
      "get": {
        "description": "Retrive all short URLS for user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "user-urls-json",
        "responses": {
          "307": {
            "description": "Temporary Redirect",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "integer"
            }
          }
        }
      },
      "delete": {
        "description": "Delete all short URLS for user",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "user-urls-json",
        "responses": {
          "202": {
            "description": "Accepted",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    },
    "/ping": {
      "get": {
        "description": "Ping Database for check connection",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "text/plain; charset=utf-8"
        ],
        "operationId": "ping",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "integer"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "dto.ShortenInput": {
      "type": "object",
      "properties": {
        "url": {
          "type": "string"
        }
      }
    }
  }
}